This is a new file that contains the definition of the `GitErrorCode` enum and the `GitOperationError` class.
The `GitErrorCode` enum defines the possible error codes that can be returned by the `git-blame` library. Each code has a descriptive name, such as `AUTH_FAILED`, `NO_REMOTE`, etc.

The `GitOperationError` class is used to wrap errors from the `git-blame` library with additional information about the error, such as the error code and the original error. This makes it easier for developers to handle errors in their applications.

Note that this file is not related to the `src/index.ts` file, which is the main entry point of the application. The `src/index.ts` file contains the definition of the `GitBlame` class and its methods.